Maze_Algorithm/pyproject.toml
[tool.poetry]
name = "micro-mouse"
version = "0.1.0"
description = ""
authors = ["<xian.chen@stud.tu-darmstadt.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.2"
pygame = "^2.5.1"
networkx = "^3.1"
matplotlib = "^3.7.3"
pyserial = "^3.5"
keyboard = "^0.13.5"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
